env_mmc: allow platform override of mmc_get_env_part()
U-Boot v2015.04 offers a weak implementation of mmc_get_env_part() only for
platforms that define CONFIG_SYS_MMC_ENV_PART.
The purpose of this function is precisely to not having to define that
constant, so this patch removes the #ifdef around the function and defines
a default value of 0 for the constant if not defined, to fulfill the weak
implementation.

On the ccimx6.c platform code, the old function is renamed to match the
upstream function name.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>