ccimx6: return valid values when not booting from known media
If booting from USB, a -1 was returned for the mmc device index and
partition where the environment is located. This resulted in error message
when trying to access the mmc device.
To avoid this error return default valid values for mmc device and index
partition where the environment is located.

These are effectively never used. The patch just avoids an error message.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>