ccimx6: Assert nRESET when coming out of power down mode.
When we detect a spurious wake we put the PMIC in power down mode, as
opposed to reset mode which is used on a normal poweroff.

When waking up from reset mode nRESET is always asserted, but when coming
out of power down mode it is not, so we need to force it for the target
to boot after an spurious wake is act upon.

https://jira.digi.com/browse/DEL-1775

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>
(cherry picked from commit 3e6410b50f3198596a10f7bd15686683025c939e)
(cherry picked from commit 8451e9da23ae8c2d63356fc86a64a364253ec2f2)