ccimx6: implement recovery for Android
Android kernel will set a flag (1 << 7) on persistent register SNVS_LPGPR
when instructed to do a "recovery" restart.
The following patch checks for this flag on the ccimx6_late_init() function
and, if set, runs the script stored in variable 'recoverycmd'.

By default, the 'recoverycmd' forcibly sets the variable 'bootargs_once'
to format Android's 'data' and 'cache' partitions. Since this variable
is flagged to be read-only, the value is lost on the next reboot (even if
a saveenv is run).

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/ADK4A-620
(cherry picked from commit f4f75d9fc154629781f0d7a21b85759d5da294d4)
Signed-off-by: Hector Palacios <hector.palacios@digi.com>

Conflicts:
	arch/arm/cpu/armv7/mx6/soc.c
	include/recovery.h