ccimx6: set PMIC i2c bus before accessing it
The board supports I2C multi-bus, so the PMIC access functions
must set the PMIC bus before trying to access PMIC addresses.
This was only done on the first function where the PMIC was
accessed but not in any other call of the PMIC API. As a consequence,
if the user changed the I2C bus, subsequent calls to the 'pmic'
command failed.

The patch moves the selection of the I2C bus to the function
'pmic_access_page()' which is called by all the functions of the
PMIC API.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-516