ccimx6: moved carrier board functions to carrier board file
A lot of the functions in the source code file ccimx6.c (reserved for code
that applies to the CC6 SOM) belong really to the carrier board domain.
In an effort to clean the code, this patch moves the following functions to
the ccimx6sbc.c file (carrier board):

 - board_phy_config() and related. These setup the Micrel PHY on the SBC.

 - get_carrierboard_version()

 - fdt_fixup_carrierboard(): sets the carrier board version on the device
   tree.

 - print_carrierboard_info()

 - checkboard(): this is called by U-Boot main code to print information
   about the carrier board.

 - ft_board_setup(): this is called by U-Boot main code to fixup the device
   tree as needed for the board.

Include headers, declaration of functions, and staticzation of functions
has been applied, as needed, but this commit does not contain any
functional change.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>