cmd_board: print warning if board version undefined
Before, this command printed the board name followed by the version:
	Carrier board: ConnectCore 6 SBC v1

Since we will now also have to print the carrier board variant we'd better
split the board name and version:
	Board: ConnectCore 6 SBC
	       Version: 1

Besides, print a WARNING message if the board version is undefined.
We want to remark this, so that customers designing their own carrier
boards, take into account if conditional code that applies to SBC should
also apply to their boards.

Finally, wrap the code around CONFIG_HAS_CARRIERBOARD_VERSION since this
is optional for a customer's carrier board.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-2160