cmd_board: on override command, override only the given field
The subcommand 'override' of commands 'board_version' and 'board_id'
writes the complete shadow register of a certain OTP word, where all the
bits, except those of the field to be overridden (version or ID) are
zero. Although this is ok for the 'prog' subcommand (since we don't want
to risk blowing additional OTP bits), it is wrong for the 'override'
subcommand, since in that case we only want to override the specific
field (version or ID), but leave the rest of the bits untouched.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>