ccimx6sbc: configure audio codec lines as GPIOs
The SGTL5000 audio chip on the SBC board does not have a reset line. If a
software reset (watchdog reset) takes place while the audio is playing a
sound (DAC enabled), the DAC remains enabled and noise can be heard until
the kernel initializes the codec on the next reboot.
Even if U-Boot is powering off the codec through a GPIO, the codec is still
able to power itself using the voltage present on the I2S bus.
By configuring the I2S bus lines as GPIOs the codec is not powered, the
DAC is disabled, and the audio cancelled.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-350