ccimx6: otf-update: skip copying buffer if data already in RAM position
OTF update was only supported for TFTP and the data packets were on the
TFTP buffer and needed to be copied to the reference address in RAM
before they were transferred to the target media.
When running OTF mechanism from a file system, the data packets are already
on the reference address in RAM, so we can skip the copying of the buffer.

This patch:
	- checks if the buffer is not NULL and the length of packet is
	  bigger than zero before trying to copy the buffer to RAM.
	- moves the copy as the first thing to do (no real functionality
	  change, but easier to understand).
	- sets chunk_len before checking if this is the last packet.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-526