tftp: set otfd->len=0 when completing the TFTP transfer
After commit b15573c22becea0a64f2f4d755ade1c9c4817815 the
on-the-fly update was writing the final chunk with the remaining
bytes in RAM plus the otfd->len bytes. This variable was not
reset after the last TFTP frame, so we were trying to write
those many bytes again.
The TFTP protocol calls the OTF flush after it has transferred
all frames, so we should set the otfd->len to zero.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-550