ccimx6: do not use local variable 'my_hwid' in functions that can be shared
Board code uses global local variable 'my_hwid' across several functions.
Some of these functions can be shared with the CC6UL platform and will be
moved to a common file.
In prevision of that move, first the functions need to not depend on the
local variable and instead have the ccimx6_hwid struct be passed as
parameter.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-559