hab: add the IMX_HAB Kconfig directive to enable the HAB API
Previous to this patch the HAB API is enabled only when SECURE_BOOT is
used. We now want to be able to use the HAB API to know the status of
the devices without enabling SECURE_BOOT.

This patch introduces the IMX_HAB directive to add granularity and be able
to use the HAB API and the secure boot support independently.

https://jira.digi.com/browse/DUB-527

Signed-off-by: Jose Diaz de Grenu de Pedro <Jose.DiazdeGrenudePedro@digi.com>