ccimx6ul: reorganize platform and common headers
The platform header includes the common header at the first line, which
means this common header cannot have #ifdef blocks that depend on constants
defined on the platform header (because they are defined later).
This patch moves things around to avoid this erros and also groups some
defines by interface.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>