nand: compile nand_device_info only for cmd_bootstream support
The i.MX6 uses the MXS NAND controller but the compilation of
cmd_bootstream.c is still not supported. The file nand_device_info.c is
only required for updating U-Boot in the flash and it currently does
not build. Fix the Makefile so that this is only built when
CONFIG_CMD_BOOTSTREAM is enabled.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>