scripts: fix inconsistent U-Boot file when generating signed images
When CONFIG_MAKE_SIGNED is defined both an unsigned and a signed U-Boot images
are generated. The nonsigned U-Boot image is generated with a non-zero CSF
pointer in its IVT table even when there is no CSF attached. This can cause
the boot procedure to fail when flashing an unencrypted U-Boot after flashing
an encrpyted one.

https://jira.digi.com/browse/DUB-569

Signed-off-by: Jose Diaz de Grenu de Pedro <Jose.DiazdeGrenudePedro@digi.com>