autoboot: do not reset environment when booting from USB
NXP BSP resets the environment to default values when it
detects that the device is being booted from USB.
Do not reset it and only print a message about booting the
manufacturing script in 'bootcmd_mfg' if the variable exists.
Otherwise, print a message warning that the variable is
empty.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>