ccimx6ul: skip MCA reset implementation if doing a bmode reset
The board_reset implementation for the CC6UL is managed by the MCA which
in turn does a power cycle of the CPU. This prevents from using the 'bmode'
command, which does a real soft-reset of the CPU to override the BOOT_CFG
strapping using SRC registers.

This commit flags the request of a bmode_reset so that we can skip the MCA
reset implementation.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-609