update_nand: add support for updating UBIFS images to NAND partitions
Digi 'update' command raw-writes files into NAND partitions.
This commit adds support to detect if a partition is UBI formatted or
a UBIFS file is being written, in order to do the update using
'ubi write' instead of doing a raw-write using 'nand write'.

The logic followed is the following:

 1. The first page of the partition is read, and the first word compared to
    UBI magic fingerprint "UBI#".

 2. If the partition is UBI-formatted (see 1) the first volume name is read.

 3. If there is no volume, one is created with the same name of the MTD
    partition.

 4. If the partition is not UBI-formatted but a file with extension *.ubifs
    is passed, a UBI volume is created.

 5. If the partition contains a UBI volume 'ubi write' is used, otherwise
    'nand write' is used (raw write).

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-604