cmd_bootstream: update code to work with U-Boot MTD framework
This porting of kobs-ng was done basing on old Digi NVRAM framework.
Change mtd functions and headers to make it build with standard U-Boot
code.

This also removes the support for getting more accurate NAND timings
from a NAND database, which was not building either.

Besides, it removes the functions for supporting MX53 which were never
used.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>