defconfig: enable SECURE_BOOT by default on ConnectCore 6 and ConectoCore 6UL
This has no effect on open devices, but on closed devices it will verify
the OS image signature before booting it.

Having this setting enabled by default allows to sign and encrypt existing
U-Boot images and uImages and to establish a secure state without the need
to recompile U-Boot.

Also, improve the description of the Kentry.

Signed-off-by: Diaz de Grenu, Jose <Jose.DiazdeGrenu@digi.com>