ccimx6sbc/ccimx6ulsbc/ccimx6ulstarter: Separate setting GPI direction from reading state.
On the i.MX6 there is a delay between configuring an internal pull-up
or pull-down pad configuration, setting it as input, and the PU/PD being
applied.

This means that you need to wait approximately 10 usecs between setting a
pin as input and reading it if you want the PU/PD to be active.

Because the check for a console enable GPIO happens very early on, before
the timers have been initialized, we cannot use udelay() to separate the
setting up of the input direction and the reading.

This commit configures the console enable GPI as input ahead of reading its
state.

https://jira.digi.com/browse/DUB-597

Signed-off-by: Alex Gonzalez <alex.gonzalez@digi.com>