ccimx6ul: fix boot problems with ConnectCore 6UL
The upstream code from NXP assumes that a CSF blob is going to be appended
to the U-Boot image when SECURE_BOOT is defined, so defines a CSF pointer
in those situations.

We separate secure U-Boot from secure OS, so the CSF configuration is done
through the CONFIG_CSF_SIZE Kconfig entry. That support was added in
cb8b34548a83bf6985f3b26424cb2fcc8d7034f0, but the existing NXP code for
setting the CSF pointer was not removed. This causes U-Boot images with
an invalid CSF pointer when CONFIG_SECURE_BOOT is enabled, which can then
fail to boot.

This patch fixes it by removing the NXP code which defined a CSF pointer
when CONFIG_SECURE_BOOT is enabled.

https://jira.digi.com/browse/DUB-624

Signed-off-by: Diaz de Grenu, Jose <Jose.DiazdeGrenu@digi.com>