sign.sh: avoid modifications on the original U-Boot image
The CST uses the same file as input and output, but the default U-Boot
image should not be modified during the signature and encryption processes.
In order to avoid any modifications, copy the U-Boot image before running
the process, and restore it at the end.

Note that the CST pointer still needs to be cleared out, because it gets
set during the compilation process (if CONFIG_CSF_SIZE is enabled).

https://jira.digi.com/browse/DEL-2733

Signed-off-by: Diaz de Grenu, Jose <Jose.DiazdeGrenu@digi.com>