ccimx6ul: avoid MCA error access messages get in the way
The U-Boot banner displays the MCA hardware and firmware version
or a "??" if the version could not be determined.
If there are errors accessing the MCA, the error messages get in
the way of the banner line, breaking it.
By calling the MCA functions first and printing the results afterwards
the line is not broken.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>