digi_common: fix help text of 'dboot' in argument [filename]
Constants were used to generate the help text of commands 'update' and
'dboot' which use a very similar syntax.
The update command however uses for the default filename the one stored
in variable $<partition>_name, whereas dboot uses the name stored on
variable $uimage or $zimage, as per CONFIG_DBOOT_DEFAULTKERNELVAR.

This commit fixes it splitting the help text in two constants, instead
of reusing the same for both commands.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-654