trustfence: add an independent rule for encrypted images
Before this patch the same rule (%-signed.imx) was being used for signed
and encrypted images. This made it impossible to build both images in the
same make call.

This patch adds a rule for encrypted image fixing that problem. Note that the
signed image is introduced as a dependency for the encrypted one to avoid
data races, as the sign/encrypt script is not thread safe.

https://jira.digi.com/browse/DUB-642

Signed-off-by: Diaz de Grenu, Jose <Jose.DiazdeGrenu@digi.com>