imx6: caam: introduce key_modifier as a parameter for data blob operations
This is required in order to use the CAAM blob encapsulation and
decapsulation operations with a secure unique key modifier, instead of the
default one.

When using the caam U-Boot command a null key modifier will be used.

For U-Boot environment encryption the md5 hash of the OTP words containing
the digi HWID (unique per module) is used as key modifier.

https://jira.digi.com/browse/DUB-652

Signed-off-by: Diaz de Grenu, Jose <Jose.DiazdeGrenu@digi.com>