fsl: caam: DMA-align cache flush and invalidation ranges for the CC6UL
The contents of these memory locations are to be used by the CAAM, which copies
to and from them using DMA. This makes it so that the ranges must be aligned to
the minimum DMA aligment.

The hardcoded value was good enough for the CC6 (which has ARCH_DMA_MINALIGN = 32)
but it is incorrect for the CC6UL (which defines that constant to 64).

Signed-off-by: Diaz de Grenu, Jose <Jose.DiazdeGrenu@digi.com>