env: enable CAAM environment encryption only on closed devices
The encryption is not safe when done in an open device, because a public
test key is used.

This also simplifies the environment recovery: if the environment
encryption is done on an open device and then that device gets closed, it
would be difficult to decrypt the environment (different keys are used).

Signed-off-by: Diaz de Grenu, Jose <Jose.DiazdeGrenu@digi.com>