scripts: sign.sh: set the CSF pointer for the signed images
When the script is used within the compilation process, the input U-Boot
image (u-boot.imx) contains a valid CSF pointer in its IVT, which is
written during compilation based on the value of CONFIG_CSF_SIZE.

However unsigned U-Boot images have a null pointer in this field (which is
correct because they don't have a CSF binary attached). In order to be able
to use this script to sign existing U-Boot images, it is necessary that the
script replaces the CSF null pointer with the location of the CSF binary
to be appended.

https://jira.digi.com/browse/DEL-2688

Signed-off-by: Diaz de Grenu, Jose <Jose.DiazdeGrenu@digi.com>