trustfence: generate u-boot-usb-signed.imx image for secure USB boot
When recovering a device via USB, the Device Configuration Data (DCD) is
copied to the On Chip RAM (OCRAM). This invalidates the U-Boot signature
(which covers the DCD) so the default signed images cannot be used.

This patch adds a new signed image with a nulled DCD pointer. This image
can be used to boot a closed device via USB once it has been initialized.

https://jira.digi.com/browse/DEL-2843

Signed-off-by: Diaz de Grenu, Jose <Jose.DiazdeGrenu@digi.com>