imx6: enable OCOTP_CTRL clock (iim_clk_enable)
This clock was disabled to save power. However this prevents the ROM from
updating the sticky bit in the OCOTP controller in order to lock the
SRK_REVOKE eFuses.

As a result, if this clock is disabled it is possible to revoke keys even in
closed devices.

This patch enables the clock so that key revocation is properly locked in
closed devices.

https://jira.digi.com/browse/DUB-668

Signed-off-by: Diaz de Grenu, Jose <Jose.DiazdeGrenu@digi.com>