ccimx6ulsbc: add support for microSD card on uSDHC2
The microSD card on the SBC board uses the same signals as the one on the
Starter Board.
This commit disregards yet the electronic switch to select between eMMC and
microSD, since it defaults to SD.

The commit also removes other uSDHC controllers initialization remaining
from the EVK source code, when this platform was created.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-672