hab: reduce verbosity of secure boot
After adding support for FDT and ramdisk authentication, there are too many
lines about authentication during boot. This patch reduces it to one line
per artifact for successful authentications. The detailed information is given
only if the validation fails.
Also reword the message shown when the validation is skipped (on open devices)

Signed-off-by: Diaz de Grenu, Jose <Jose.DiazdeGrenu@digi.com>