trustfence: update: allow to reuse existing DEK
The 'trustfence update' allows you to flash an encrypted U-Boot in the target.
However, this command needs the Data Encryption Key in plain text, which is
inconvenient.

This patch allows to reuse the secure DEK already in the target. That will be
done when the last parameter (plain DEK file) is skipped. Obviously the
encrypted U-Boot image being flashed must be encrypted with the same DEK that
is already in the target.

https://jira.digi.com/browse/DUB-643

Signed-off-by: Diaz de Grenu, Jose <Jose.DiazdeGrenu@digi.com>