ubi: allow to top-limit reserved blocks for bad block handling
The UBI subsystem takes a very conservative approach when
calculating the blocks it must reserve for bad block handling
and calculates them basing on the total size of the NAND
rather than basing on the size of the MTD partition.
For small partitions, the number of blocks thus reserved may
be too large in proportion to the partition size.

For example, on a NAND with total size of 256MiB, a partition
of 14MiB reserved 40 blocks (5MiB), which is more than 35% of
the partition.

This commit allows to define a max percentage limit per
partition on the number of blocks to reserve to account for
potential bad blocks.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-2840