ccimx6ulstarter/sbc: use a variable to determine the boot partition
The $loadscript run by the default boot command was hardcoding the MTD
partition where to read from (CONFIG_LINUX_PARTITION).
Create a new variable $mtdbootpart to hold the partition name where to
boot from. This way, the partition to boot from can be selected by a user
without needing to change the default boot command.

In the $loadscript, if the variable does not exist, set it to
CONFIG_LINUX_PARTITION by default.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-698