ccimx6/ccimx6ul: simplify $recoverycmd by calling default $bootcmd
We want the $recoverycmd to run similarly than default $bootcmd, that is
to read out and run a script from the recovery partition.
The only thing that it must then do is to set the variable that represents
the recovery partition and call the default boot command.

This also guarantees that the script in the recovery partition is
authenticated by TrustFence.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-698