ccimx6ulstarter/sbc: add mtd partitions to all cmdline
When booting from TFTP/NFS or MMC, the kernel command line
does not contain the MTD partitions, this makes the NAND
partitions unusable when booting from NFS, for example.

This patch adds the MTD partition table unconditionally
to all CC6UL bootargs methods, given that the NAND is
an always present chip.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>