ccimx6: run recovery command by setting 'boot_recovery'
Instead of running directly the recovery command, flag the boot_recovery
variable, so the recovery command is selected instead of the normal
boot command at the end of the boot process when doing the autoboot.

This fixes a corner case that happens when the recovery command fails
somehow (for example if the recovery partion has not been flashed)
and then it tries to run the standard bootcmd, but at this point the
MMC partition to boot from has also been changed so it fails as well.

https://jira.digi.com/browse/DUB-701

Signed-off-by: Javier Viguera <javier.viguera@digi.com>