ccimx6/6ul: remove bootargs_recovery
When booting in recovery mode ($boot_recovery=yes) the dboot
command will run the script on $bootargs_recovery if it exists
or the $bootargs_<src>_<os> otherwise.
At the moment however, the recovery ramdisk is able to work
with the default command line and does not need any special
arguments for recovery.
A user may still set this variable to change the default
arguments if needed.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-708