trustfence: fix clock management on dek_blob command
The CAAM clocks need to be enabled before attempting to read the
ORSR_JR0 register. This was causing the device to hang when generating DEK blobs
if these clocks were disabled (for example, after using the authenticate_image
function).

Enable the clocks before reading the register and disable them afterwards.

https://jira.digi.com/browse/DUB-720

Signed-off-by: Diaz de Grenu, Jose <Jose.DiazdeGrenu@digi.com>