ccimx6/ccimx6ul: move dynamic variable generation to platform hook
Digi platforms generate some environment variables upon start. These are
not generated when the environment is reset, though.
With the recent addition of a platform-specific hook to set the default
environment, we can now move the generation of such variables into this
hook.

For SOM-specific variables, a function is created inside the SOM code.

Although this mechanism would allow to remove the write-once flag for
these variables, we'd rather keep it, to prevent accidental modification.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>