ccimx6ulstarter/sbc: move redundant environment in 1GB variant
For the 1GB variant, place the redundant copy half way across the
environment partition, to account for NAND flash chips with bigger erase
block size than 128K and to give enough blocks to each copy for bad block
management.

For the 256M variant, leave the redundant copy where it was when the
variant was released, to avoid backwards-compatibility problems.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-737