env_nand: add support for dynamic location of environment
The current implementation to store the environment requires to define an
offset for the primary environment and another for the redundant copy.
To account for bad blocks, CONFIG_ENV_RANGE must defined which establishes
a range of sectors (starting from each copy offset) where each copy can
attempt to write to.

This mechanism has some disadvantages:
- The amount of blocks to be used in case of bad blocks for each copy is
  half of the full number of blocks reserved for the environment.
- For platforms that have released with fixed offsets cannot grow the area
  reserved for the environment, if so required.

The following commit implements a new mechanism to fix these two issues.
Before each save operation, a function will determine for each copy of the
environment, if its block is good. If it is not, it will be relocated to
the next good block within the area determined by CONFIG_ENV_RANGE.

The new mechanism can be enabled by defining CONFIG_DYNAMIC_ENV_LOCATION.
Requisites are:
- CONFIG_ENV_OFFSET should point to the offset where the environment area
  starts. This will also be set as the initial offset for the primary
  copy of the environment.
- CONFIG_ENV_RANGE should be set to the size of the environment area. It
  should be big enough to fit two copies of the environment (if redundant
  copy is enabled) and empty blocks for bad block management.
- CONFIG_ENV_OFFSET_REDUND should point to the initial offset where the
  redundant copy of the environment starts. Typically this would be the
  next erase block, but since the erase block may be different depending
  on the NAND geometry it can be left undefined and the config_fallbacks.h
  will set it to the same offset than the primary. The code will then later
  locate it at the next good sector after the primary copy.
- CONFIG_ENV_SIZE_REDUND can be left underfined and it will fall back to
  the same size than the primary copy.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-741