trustfence: move the encryption key to the 'safe' partition
The TF encryption key was located after the environment redundant copy.
If CONFIG_DYNAMIC_ENV_LOCATION is enabled and bad blocks are
present in the 'environment' partition, the environment copies will overwrite
the TF key.

This mechanism checks for valid TF key on the 'environment' partition.
If there is a valid key and the 'safe' partition is empty, it copies the TF key
to the 'safe' partition, where the new TF implementation expects it.

https://jira.digi.com/browse/DUB-739

Signed-off-by: Isaac Hermida <isaac.hermida@digi.com>