digi: fixup DT with boolean property 'digi,uboot,dynamic-env' if enabled
To maintain backwards compatibility in DEY, the U-Boot that supports the
dynamic environment location will add the boolean property
'digi,uboot,dynamic-env' to the device tree, if the mechanism is enabled.

While on it, rename the function that fixes up the DT to add the U-Boot
version into more generic fdt_fixup_uboot_info().

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DEL-4256