mtd: nand: mxs: move custom initialization to post identification hook
Certain setup of the NAND controller was being done in the scan_bbt() hook
since it is one of the first operations carried out after identification
of the NAND. If we want to skip the intial bad block table scanning, the
initialization of the controller would not be done because of this, so this
commit moves this initialization to a new post-identification hook that is
called right after nand_scan_tail().
A weak stub implementation is provided for platforms that don't need it.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-755