ccimx6qp: add support to ConnectCore 6+ variants re-using ccimx6 platform
ConnectCore 6+ is a SOM based on ConnectCore 6, with the i.MX6QP or i.MX6DP
CPU and a different Wi-Fi chip. The SOM has a different part number so this
commit creates a new array to manage the cc6p variants and their DDR3
calibration values.

The first variant (0x02) has 2GB of DDR3 and its calibration values are,
for the moment, copied from its CC6 sibling (variant 0x12).

Signed-off-by: Hector Palacios <hector.palacios@digi.com>

https://jira.digi.com/browse/DUB-762