ccimx6: move DDR3 timings and calibration into an i.MX6Q folder
The i.MX6QP (QuadPlus) has some minor differences in some registers which
will prevent us from using the same configuration values than the ones used
on the ConnectCore 6, so we need to place these files under different
folders. the

Although the calibration files won't have the same names (because the
platform name is part of the filename) this commit also moves them, so that
all CPU-specific files are together.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>