ccimx6: modify the board name for CC6PLUS on-the-fly
U-Boot reuses the platform code and folders from the ConnectCore 6 for
the ConnectCore 6 QuadPlus, but we want to change the CONFIG_SYS_BOARD
to match the QuadPlus platform name 'ccimx6qpsbc' so that it matches
the filenames expected by Digi Embedded Yocto.

This changes the variable 'board_name' and other variables that used
CONFIG_SYS_BOARD as part of their default value, like $uimage, $fdt_file.

Also make sure the prefix of $fdt_file corresponds to the iMX6 family.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>