ccimx8x: global rename from 'ccimx8qxp' to 'ccimx8x'
The family i.MX8X contains three CPU variants:
 - QuadXPlus (QXP)
 - DualXPlus (DXP)
 - DualX (DX)

Let's rename from the specific CPU variant to the family so that we turn
the file names and platform names CPU-variant agnostic.

This includes removing the 'imx8qxp-' prefix from the U-Boot device tree
file.

While on it, change the underscores in the 'compatible' string to dashes
and rename the pinctrl node from 'mek' to 'ccimx8x'.

Signed-off-by: Hector Palacios <hector.palacios@digi.com>