Revert "arm: Show cache warnings in U-Boot proper only"
This commit causes cache misalignment warnings to always be printed on non-SPL
images, regardless of being in debug mode or not. By reverting it, the warnings
will appear exclusively in debug mode.

This warning seems to be harmless, as most functions that use the
check_cache_range() function don't check its return code. For the one function
that does, its implementation was different in v2015.04 (meaning the check
never failed) and the warning never appeared.

This reverts commit bcc53bf095893fbdae531a9a7b5d4ef4a125a7fc.

https://jira.digi.com/browse/DUB-845