ccimx8x: add board_version and board_id support
Due to constraints in the OTP words availability, board_id and
board_version are defined as write-once environment variables.
These variables are protected against regular setting, resetting or
default setting once they are set for the first time.

In order to modify them after the first time, a forced action should be
issued adding '-f' to the regular command.

Commands board_id and board_version will not be available in this platform.

https://jira.digi.com/browse/DUB-824

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>