ccimx8x: set mac addresses from fuses
This commit adds the support to set the mac addresses from the
fuses.

On the CC8X platform, the MAC addresses are stored in the HWID in the form of
MAC pool and MAC base.

The MAC addresses are formed combining the 3 upper bytes, provided by the used
MAC pool, and the 3 lower bytes that are programmed directly into the HWID.

The MACs for the different interfaces (eth, eth1, wlan, bt) are consecutive and
the number of macs is based on the variant.

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>