ccimx6ulsbc: add video and logo support
This commit adds video and logo support to the ccimx6ulsbc platform. The
feature is disabled by default and can be enabled by selecting the config
options: CONFIG_VIDEO and CONFIG_CFB_CONSOLE.

The format of the logo is a BMP file that can be specified with the 'logosrc'
environment variable. The file can be stored in any media supported by the
platform like the on board flash, the sdcard or a usb stick. The format for
logosrc variable is like the arguments of the dboot command. For instance, to
load the logo from a file called mylogo.bmp on the first partition of the
uSD card, the variable should be defined as follows: "linux mmc 0 mylogo.bmp".

Signed-off-by: Pedro Perez de Heredia <pedro.perez@digi.com>
Signed-off-by: Hector Palacios <hector.palacios@digi.com>