helper: otf: account for possible offsets
If the write handler is not able to write the complete OTF chunk of data, it
may keep some of it and set the 'offset' variable to deal with this in the
next OTF chunk. In that case the read handler should not overwrite the range
[0, otfd.offset), but only fill [otfd.offset, CONFIG_OTF_CHUNK).

This was not being taken into account by the 'write_file_fs_otf' function.

Note that this cannot happen when using raw images, as in those cases the
data is read and written in chunks of the storage media block size which
is a divisor of CONFIG_OTF_CHUNK. However, it can happen when dealing with
sparse images.

Also correct comment to clarify this requirement.

https://jira.digi.com/browse/DUB-868

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>