helper_mmc: ensure reads use an aligned buffer
When using the 'load' command to read a file from mmc, the memory address
and the file offset must be congruent module ARCH_DMA_MINALIGN. Otherwise
warnings may appear and the operation may fail.

That is, when running:

* load mmc <dev:part> <loadaddr> <file_name> <size> <file_offset>

The following condition must hold:

* loadaddr % ARCH_DMA_MINALIGN == file_offset % ARCH_DMA_MINALIGN

To prevent these errors, move pending data to the right so that its end is
DMA-aligned. As new data is read right after the pending data, and the file
offset used is always an mmc block multiple (which is a multiple of
ARCH_DMA_MINALIGN itself), this ensures that the previous condition is always
met.

Note that this can only happen when using sparse images, so this workaround has
been purposely implemented in sparse-flashing specific code, leaving the
general OTF code unaware of this.

https://jira.digi.com/browse/DUB-868

Signed-off-by: Jose Diaz de Grenu <Jose.DiazdeGrenu@digi.com>