fastboot: only modify the default bootargs when CMD_BOOTA is enabled
When the bootcmd environment variable is deleted, the target won't boot
automatically when reset. However, fastboot has logic that changes the variable
automatically if it isn't set, inserting the Android-specific "boota" command
instead. This command isn't even enabled in our U-Boot yet, causing the command
to default to "bootaux" and triggering a core dump due to the lack of Cortex M4
demos in our default images.

For now, make fastboot perform this action only if CMD_BOOTA is enabled.

Signed-off-by: Gabriel Valcazar <gabriel.valcazar@digi.com>