trustfence: fix GPIO console enable
Read the GPIO used to enable the console in a point where the DM GPIO driver
is already initialized.

The GPIO will be read in the same moment of the initialization that the
passphrase would be check to enable the console.

https://jira.digi.com/browse/DUB-870

Signed-off-by: Gonzalo Ruiz <Gonzalo.Ruiz@digi.com>
(cherry picked from commit 8ee0ff7f9ebfde8706934d86eb88ffb76daf770c)
Signed-off-by: Hector Palacios <hector.palacios@digi.com>