fastboot: allow to flash 'boot' partition with sparse image
NXP's implementation assumes the boot partition is for an Android boot
image type (to be booted with 'boota' command), and thus it's filtering
out that partition from the support to program sparse images.

This is not our case, we use sparse images for all partitions (boot,
recovery, vendor and system), so remove the filter to allow flashing
sparse boot images.

https://jira.digi.com/browse/ADK4A-1359

Signed-off-by: Javier Viguera <javier.viguera@digi.com>